Fixed error caused by incorrect use of lolcat-c
When I installed this project, I noticed that instead of what I expected, I got a long prompt that seemed to have the lolcat help menu embedded in it. I fixed this bug and decided to put it here. This is the copy of .bash_prompt I'm actually using, so it will have my personal changes as well.